refactor: key 생성 기능 클래스로 분리
