+++ src/test/java/com/knu/ddip/location/infrastructure/repositoroy/LocationWriterImplTest.java
@@ -4,7 +4,6 @@ import com.knu.ddip.config.IntegrationTestConfig;
 import com.knu.ddip.config.MySQLTestContainerConfig;
 import com.knu.ddip.config.RedisTestContainerConfig;
 import com.knu.ddip.config.TestEnvironmentConfig;
-import com.knu.ddip.location.exception.LocationNotFoundException;
 import com.knu.ddip.location.infrastructure.entity.LocationEntity;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -16,12 +15,10 @@ import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.UUID;
 import java.util.stream.Collectors;
 
+import static com.knu.ddip.location.application.util.LocationKeyFactory.*;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 @Transactional
 @SpringBootTest
@@ -92,88 +89,31 @@ class LocationWriterImplTest {
     }
 
     @Test
-    void validateLocationByValidCellIdTest() {
+    void cleanupExpiredUserLocationsTest() {
         // given
-        String validCellId = "validCellId";
-        locationJpaRepository.save(LocationEntity.create(validCellId));
+        long now = System.currentTimeMillis();
 
-        // when // then
-        assertDoesNotThrow(() -> locationReader.validateLocationByCellId(validCellId));
-    }
+        String cellId = "cellId";
+        String cellUsersKey = "cell:" + cellId + ":users";
+        String cellExpiryKey = "cell:" + cellId + ":expiry";
 
-    @Test
-    void validateLocationByInvalidCellIdTest() {
-        // given
-        String validCellId = "invalidCellId";
+        // 만료된 위치
+        String userId = "userId";
+        long expireAt = now - (2 * 3600 * 1000L);
+        redisTemplate.opsForSet().add(cellUsersKey, userId);
+        redisTemplate.opsForZSet().add(cellExpiryKey, userId, (double) expireAt);
 
-        // when // then
-        assertThatThrownBy(() -> locationReader.validateLocationByCellId(validCellId))
-                .isInstanceOf(LocationNotFoundException.class)
-                .hasMessage("위치를 찾을 수 없습니다.");
-    }
-
-    @Test
-    void findAllLocationsByCellIdInTest() {
-        // given
-        List<String> cellIds = List.of(
-                "findAllLocationsByCellIdInTest1",
-                "findAllLocationsByCellIdInTest2"
-        );
-        List<LocationEntity> locations = cellIds.stream()
-                .map(LocationEntity::create)
-                .collect(Collectors.toList());
-        locationJpaRepository.saveAll(locations);
+        // 만료되지 않은 위치
+        String userId2 = "userId2";
+        long expireAt2 = now + 3600 * 1000L;
+        redisTemplate.opsForSet().add(cellUsersKey, userId2);
+        redisTemplate.opsForZSet().add(cellExpiryKey, userId2, (double) expireAt2);
 
         // when
-        List<String> findCellIds = locationReader.findAllLocationsByCellIdIn(cellIds);
+        locationWriter.cleanupExpiredUserLocations(now);
 
         // then
-        assertThat(findCellIds).hasSize(2)
-                .containsAll(cellIds);
+        assertThat(redisTemplate.opsForSet().isMember(cellUsersKey, userId)).isFalse();
+        assertThat(redisTemplate.opsForZSet().score(cellExpiryKey, userId)).isNull();
     }
-
-    @Test
-    void findUserIdsByCellIdsTest() {
-        // 여 테스트 작성
-        // given
-        List<String> cellIds = List.of(
-                "findUserIdsByCellIdsTest1",
-                "findUserIdsByCellIdsTest2"
-        );
-        List<LocationEntity> locations = cellIds.stream()
-                .map(LocationEntity::create)
-                .collect(Collectors.toList());
-        locationJpaRepository.saveAll(locations);
-
-        List<String> userIds = List.of(
-                UUID.randomUUID().toString(),
-                UUID.randomUUID().toString()
-        );
-
-        for (int i = 0; i < 2; i++) {
-            locationWriter.saveUserIdByCellIdAtomic(cellIds.get(i), false, userIds.get(i));
-        }
-        // 포함되지 않는 셀, 유저 데이터
-        locationWriter.saveUserIdByCellIdAtomic("notIncludedCellId", true, UUID.randomUUID().toString());
-
-        // when
-        List<String> findCellIds = locationReader.findUserIdsByCellIds(cellIds);
-
-        // then
-        assertThat(findCellIds).hasSize(2)
-                .containsAll(userIds);
-    }
-
-    private String createUserIdKey(String encodedUserId) {
-        return "user:" + encodedUserId;
-    }
-
-    private String createCellIdUsersKey(String cellId) {
-        return "cell:" + cellId + ":users";
-    }
-
-    private String createCellIdExpiriesKey(String cellId) {
-        return "cell:" + cellId + ":expiry";
-    }
-
 }
